requirements.txt
streamlit==1.52.1
google-generativeai==0.8.5
firebase-admin==7.1.0
pandas==2.3.3
tzdata; platform_system=="Windows"
chromadb==0.6.1.0
